refactor: group platform services under core
