+++ lib/core/platform/media/camera_service.dart
