+++ lib/core/platform/notifications/notifications_module.dart
