+++ lib/core/platform/share/share_module.dart
