+++ lib/features/bookmarks/presentation/bloc/bookmark_form/bookmark_form_bloc.dart
@@ -4,8 +4,8 @@ import 'package:injectable/injectable.dart';
 
 import '../../../../../core/analytics/analytics_service.dart';
 import '../../../../../core/domain/result.dart';
-import '../../../../../core/media/image_picker_service.dart';
-import '../../../../../core/permissions/permission_service.dart';
+import '../../../../../core/platform/media/image_picker_service.dart';
+import '../../../../../core/platform/permissions/permission_service.dart';
 import '../../../domain/usecases/create_bookmark.dart';
 import '../../../domain/usecases/get_bookmark.dart';
 import '../../../domain/usecases/update_bookmark.dart';
