+++ lib/features/bookmarks/presentation/widgets/bookmark_detail_widgets.dart
@@ -11,8 +11,8 @@ import '../../../../core/analytics/analytics_service.dart';
 import '../../../../core/di/injection.dart';
 import '../../../../core/extensions/build_context_extensions.dart';
 import '../../../../core/extensions/future_extensions.dart';
-import '../../../../core/media/video_player_service.dart';
-import '../../../../core/share/share_service.dart';
+import '../../../../core/platform/media/video_player_service.dart';
+import '../../../../core/platform/share/share_service.dart';
 import '../../../../ui/animation/widget_animations.dart';
 import '../../../../ui/theme/app_radius.dart';
 import '../../../../ui/theme/app_spacing.dart';
