+++ lib/features/bookmarks/presentation/widgets/bookmark_video_attachment_preview.dart
@@ -5,7 +5,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../../core/di/injection.dart';
 import '../../../../core/extensions/build_context_extensions.dart';
-import '../../../../core/media/video_player_service.dart';
+import '../../../../core/platform/media/video_player_service.dart';
 import '../../../../ui/theme/app_icon_size.dart';
 import '../../../../ui/theme/app_radius.dart';
 import '../../../../ui/theme/app_spacing.dart';
