+++ lib/ui/widgets/app_video_player.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:video_player/video_player.dart';
 
-import '../../core/media/video_player_service.dart';
+import '../../core/platform/media/video_player_service.dart';
 import '../animation/app_durations.dart';
 import '../theme/app_icon_size.dart';
 import '../theme/app_spacing.dart';
