+++ lib/ui/widgets/app_video_player_controls.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:video_player/video_player.dart';
 
-import '../../core/media/video_player_service.dart';
+import '../../core/platform/media/video_player_service.dart';
 import '../theme/app_spacing.dart';
 import 'app_video_player_controls_bar.dart';
