+++ lib/ui/widgets/app_video_player_controls_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
 
-import '../../core/media/video_player_service.dart';
+import '../../core/platform/media/video_player_service.dart';
 import '../theme/app_spacing.dart';
 
 class AppVideoPlayerControlsBar extends StatelessWidget {
