+++ test/test_utils/mocks.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter_starter_template/core/analytics/analytics_service.dart';
-import 'package:flutter_starter_template/core/media/image_picker_service.dart';
-import 'package:flutter_starter_template/core/permissions/permission_service.dart';
+import 'package:flutter_starter_template/core/platform/media/image_picker_service.dart';
+import 'package:flutter_starter_template/core/platform/permissions/permission_service.dart';
 import 'package:flutter_starter_template/features/auth/domain/repositories/auth_repository.dart';
 import 'package:flutter_starter_template/features/auth/domain/usecases/delete_account.dart';
 import 'package:flutter_starter_template/features/auth/domain/usecases/register.dart';
