+++ test/ui/widgets/app_video_player_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_starter_template/core/media/video_player_service.dart';
+import 'package:flutter_starter_template/core/platform/media/video_player_service.dart';
 import 'package:flutter_starter_template/ui/widgets/app_video_player.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
